Define string size limits; allow single-valued string.measure
In this PR we take a step towards defining the meaning of cursor values,
by limiting them to [0,2^31].  This also implicitly restricts the
maximum number of codepoints in a string, allowing us to avoid thinking
about overflow when counting codepoints.

When working with strings as encoded byte sequences, limit to 2^31-1
bytes.  This avoids some overflow hazards and also allows
`string.measure` to return -1 to indicate failure, instead of returning
two values.

Fixes #4.