Add stringviews, concat, and slice
As discussed in https://github.com/wingo/stringrefs/issues/31, this
update separates out two halves of the proposal: a basic "stringref"
facility and an encoding-specific "views" facility.  The basic
stringrefs work with whole strings.  The views can denote positions in
strings, measure encoded sizes for substrings, and take slices.  The
"iterator" view is actually stateful and closer to the original opaque
cursor design.

Also update scope to include concatenation and slice.

Update goals to reflect that we win when we can get a good solution for
many languages, not just C++ -- so no need to mention it specifically.